refactor read cell helper file
