+++ src/domain/helpers/readCell.ts
